edit payment info model

+++ src/main/java/com/ecoffee/ecoffee_service/services/paymentService/models/PaymentCard.java
@@ -24,9 +24,9 @@ public class PaymentCard {
     private String last_4 = "";
 
     @Builder.Default
-    private int exp_month = 0;
+    private long exp_month = 0;
 
     @Builder.Default
-    private int exp_year = 0;
+    private long exp_year = 0;
 
 }
